requirements.txt
flask
pandas
scikit-learn
numpy
supabase
python-dotenv
